requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
et-xmlfile==1.1.0
exceptiongroup==1.2.0
greenlet==3.0.3
idna==3.6
iniconfig==2.0.0
numpy==1.26.4
opencv-python==4.9.0.80
openpyxl==3.1.2
packaging==24.0
pandas==2.2.1
pillow==10.2.0
playwright==1.42.0
pluggy==1.4.0
pyee==11.0.1
pyperclip==1.8.2
pytest==8.1.1
pytest-base-url==2.1.0
pytest-playwright==0.4.4
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.1
requests==2.31.0
six==1.16.0
text-unidecode==1.3
tk==0.1.0
tomli==2.0.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
